+++ PSMTabBar/PSMUnifiedTabStyle.m
@@ -48,17 +48,17 @@
 #pragma mark -
 #pragma mark Control Specific
 
-- (void)setLeftMarginForTabBarControl:(float)margin
+- (void)setLeftMarginForTabBarControl:(CGFloat)margin
 {
 	leftMargin = margin;
 }
 
-- (float)leftMarginForTabBarControl
+- (CGFloat)leftMarginForTabBarControl
 {
     return leftMargin;
 }
 
-- (float)rightMarginForTabBarControl
+- (CGFloat)rightMarginForTabBarControl
 {
     return 24.0f;
 }
@@ -143,7 +143,7 @@
         return NSZeroRect;
     }
     
-    float countWidth = [[self attributedObjectCountValueForTabCell:cell] size].width;
+    CGFloat countWidth = [[self attributedObjectCountValueForTabCell:cell] size].width;
     countWidth += (2 * kPSMUnifiedObjectCounterRadius - 6.0);
     if(countWidth < kPSMUnifiedCounterMinWidth)
         countWidth = kPSMUnifiedCounterMinWidth;
@@ -160,9 +160,9 @@
 }
 
 
-- (float)minimumWidthOfTabCell:(PSMTabBarCell *)cell
+- (CGFloat)minimumWidthOfTabCell:(PSMTabBarCell *)cell
 {
-    float resultWidth = 0.0;
+    CGFloat resultWidth = 0.0;
     
     // left margin
     resultWidth = MARGIN_X;
@@ -192,9 +192,9 @@
     return ceil(resultWidth);
 }
 
-- (float)desiredWidthOfTabCell:(PSMTabBarCell *)cell
+- (CGFloat)desiredWidthOfTabCell:(PSMTabBarCell *)cell
 {
-    float resultWidth = 0.0;
+    CGFloat resultWidth = 0.0;
     
     // left margin
     resultWidth = MARGIN_X;
@@ -285,7 +285,7 @@
         aRect.size.height+=0.5;
         
         // frame
-		float radius = MIN(6.0, 0.5f * MIN(NSWidth(aRect), NSHeight(aRect)));
+		CGFloat radius = MIN(6.0, 0.5f * MIN(NSWidth(aRect), NSHeight(aRect)));
 		NSRect rect = NSInsetRect(aRect, radius, radius);
 		
 		[bezier appendBezierPathWithArcWithCenter:NSMakePoint(NSMinX(rect), NSMinY(rect)) radius:radius startAngle:180.0 endAngle:270.0];
@@ -363,7 +363,7 @@
 - (void)drawInteriorWithTabCell:(PSMTabBarCell *)cell inView:(NSView*)controlView
 {
     NSRect cellFrame = [cell frame];
-    float labelPosition = cellFrame.origin.x + MARGIN_X;
+    CGFloat labelPosition = cellFrame.origin.x + MARGIN_X;
     
     // close button
     if ([cell hasCloseButton] && ![cell isCloseButtonSuppressed]) {
